backend/requirements.txt
# Core Django - Latest Stable
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5

# Database - PostgreSQL with Connection Pooling
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Cache & Performance - Redis Integration (OPCIONAL)
django-redis==5.4.0
redis==5.0.1

# Security - Enterprise Grade
django-ratelimit==4.1.0
django-axes==6.1.1
cryptography==42.0.0
argon2-cffi==23.1.0
bcrypt==4.1.2

# Authentication & Authorization
PyJWT==2.8.0
python-decouple==3.8

# Validation
django-phonenumber-field==7.3.0
phonenumbers==8.13.27

# Image Processing - Compatible with Python 3.13
Pillow>=10.3.0

# PDF Generation - Professional
reportlab==4.0.9

# Email - Advanced SMTP
django-anymail==10.2
premailer==3.10.0

# File Upload - Cloud Storage
django-storages==1.14.2
boto3==1.34.28

# API Documentation - Enhanced
drf-spectacular==0.27.0
drf-spectacular-sidecar==2024.1.1

# Monitoring & Logging - Production Ready
sentry-sdk==1.40.0
python-json-logger==2.0.7
django-silk==5.0.4

# Task Queue - Celery Optimized
celery==5.3.6
celery[redis]==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
flower==2.0.1

# Utilities - Enhanced
python-dateutil==2.8.2
pytz==2023.3
python-slugify==8.0.4
python-magic==0.4.27
bleach==6.1.0
gunicorn==21.2.0
whitenoise==6.6.0

# Testing - Comprehensive Suite
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0
factory-boy==3.3.0
faker==22.6.0
coverage==7.4.1

# Development Tools
django-debug-toolbar==4.3.0
django-extensions==3.2.3
ipython==8.20.0
black==24.1.1
flake8==7.0.0
isort==5.13.2
pre-commit==3.6.0

# Data Export & Analysis
openpyxl==3.1.2
pandas==2.2.0
numpy==1.26.3

# API Clients & Integrations
requests==2.31.0
httpx==0.26.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Monitoring
sentry-sdk==1.40.0

# Development
django-debug-toolbar==4.2.0
ipython==8.20.0
